Create aggregation new
This has been set up to try to aggregate with the new variables. It sets the working directory as a folder with .xlsx files for each game this year (the first one being Lady of our Lake). I've ignored the "normalize" portion for now so we can do post-agg analysis on some of these (especially the rate stats OREB%, FG%, etc. that don't require normalization).

It doesn't seem to be working well on mine but perhaps you can test and help me debug.

Thanks for all of your work on this.